requirements.txt
Flask>=3.0.0
tensorflow-cpu>=2.17.0
tf-keras>=2.17.0
opencv-python>=4.10.0
numpy>=1.26.0
Pillow>=10.0.0